Move correlated observations workflow
