+++ semeio/jobs/scripts/spearman_correlation.py
@@ -4,8 +4,10 @@ from ert_data.measured import MeasuredData
 from ert_shared.libres_facade import LibresFacade
 from ert_shared.plugins.plugin_manager import hook_implementation
 from semeio.communication import SemeioScript
-from semeio.jobs.correlated_observations_scaling.exceptions import EmptyDatasetException
-from semeio.jobs.scripts.correlated_observations_scaling import (
+from semeio.workflows.correlated_observations_scaling.exceptions import (
+    EmptyDatasetException,
+)
+from semeio.workflows.correlated_observations_scaling.cos import (
     CorrelatedObservationsScalingJob,
 )
 from semeio.jobs.spearman_correlation_job.job import spearman_job
