+++ semeio/workflows/correlated_observations_scaling/__init__.py
@@ -1,0 +1,1 @@
+from semeio.workflows.correlated_observations_scaling.job import ScalingJob  # noqa
