+++ semeio/workflows/correlated_observations_scaling/cos.py
@@ -7,9 +7,9 @@ from ert_data.measured import MeasuredData
 from ert_shared.libres_facade import LibresFacade
 from ert_shared.plugins.plugin_manager import hook_implementation
 from semeio.communication import SemeioScript
-from semeio.jobs.correlated_observations_scaling import job_config
-from semeio.jobs.correlated_observations_scaling.job import ScalingJob
-from semeio.jobs.correlated_observations_scaling.obs_utils import keys_with_data
+from semeio.workflows.correlated_observations_scaling import job_config
+from semeio.workflows.correlated_observations_scaling import ScalingJob
+from semeio.workflows.correlated_observations_scaling.obs_utils import keys_with_data
 
 
 class CorrelatedObservationsScalingJob(
