+++ semeio/workflows/correlated_observations_scaling/exceptions.py
@@ -4,7 +4,7 @@ class EmptyDatasetException(ValueError):
 
 class ValidationError(ValueError):
     def __init__(self, message, errors):
-        super(ValidationError, self).__init__(message)
+        super().__init__(message)
         self.errors = errors
         self.message = message
