+++ tests/jobs/misfit_preprocessor/test_integration.py
@@ -7,7 +7,9 @@ from res.enkf import EnKFMain, ResConfig
 
 import semeio
 import semeio.jobs.scripts.misfit_preprocessor as misfit_preprocessor
-from semeio.jobs.correlated_observations_scaling.exceptions import EmptyDatasetException
+from semeio.workflows.correlated_observations_scaling.exceptions import (
+    EmptyDatasetException,
+)
 from tests.jobs.correlated_observations_scaling.conftest import TEST_DATA_DIR
 
 from unittest.mock import Mock
